src/test/java/net/dgazdag/recipe/controllers/IndexControllerTest.java
package net.dgazdag.recipe.controllers;

import net.dgazdag.recipe.domain.Recipe;
import net.dgazdag.recipe.services.RecipeService;
import org.junit.Before;
import org.junit.Test;
import org.mockito.ArgumentCaptor;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import org.springframework.ui.Model;

import java.util.HashSet;
import java.util.Set;

import static org.junit.Assert.*;
import static org.mockito.Mockito.*;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;

public class IndexControllerTest
{

  IndexController indexController;

  @Mock
  Model model;

  @Mock
  RecipeService recipeService;

  @Before
  public void setUp() throws Exception
  {
    MockitoAnnotations.initMocks(this);
    indexController = new IndexController(recipeService);
  }

  @Test
  public void testMockMvc() throws Exception
  {
    MockMvc mockMvc = MockMvcBuilders.standaloneSetup(indexController).build();

    mockMvc.perform(get("/"))
        .andExpect(status().isOk())
        .andExpect(view().name("index"));
  }

  @Test
  public void getIndexPage() throws Exception
  {
    // given
    Set<Recipe> recipesData = new HashSet<>();
    recipesData.add(new Recipe());

    Recipe recipe = new Recipe();
    recipe.setId(1L);
    recipesData.add(recipe);

    when(recipeService.getRecipes()).thenReturn(recipesData);

    ArgumentCaptor<Set<Recipe>> argumentCaptor = ArgumentCaptor.forClass(Set.class);

    // when
    String viewName = indexController.getIndexPage(model);

    // then
    assertEquals(viewName, "index");
    verify(recipeService, times(1)).getRecipes();
    verify(model, times(1)).addAttribute(eq("recipes"), argumentCaptor.capture());
    Set<Recipe> setInController = argumentCaptor.getValue();
    assertEquals(2, setInController.size());
  }

}